+++ CHANGELOG.md
@@ -27,6 +27,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2023-11-05
 
 ### ADDED
+- Method to get the assembly title.
+- Method to get the assembly version.
+- About property.
+
+---
+
+## [Unreleased] - 2023-11-05
+
+### ADDED
 - Additional information in AssemblyInfo.
 
 ---